api/python/requirements-test.txt
# Test dependencies for PaddleOCR-json Python API
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pillow>=9.0.0